refactor(storybook): split Form stories by adapter into Conform/RHF folders
Reorganises every Form story into apps/storybook/stories/features/form/:
  - conform/{basic,fields,advanced,dialog,stepper}.stories.tsx
  - react-hook-form/{basic,fields,advanced,dialog,stepper}.stories.tsx
  - _shared.tsx (extracted story bodies + schemas; not picked up by the
    Storybook glob)

Also fixes the zod required-field convention in _shared.tsx so both
adapters surface the same message. Conform's parseWithZod strips empty
inputs to undefined (zod v4 default error "Invalid input"); the new
convention pairs z.string({ error: msg }) with .min(1, msg) so the
custom message fires on both code paths.
