+++ apps/storybook/stories/features/form/react-hook-form/advanced.stories.tsx
@@ -1,0 +1,20 @@
+import type { Meta, StoryObj } from 'storybook-react-rsbuild'
+import { RHFAdapter } from '@datum-cloud/datum-ui/form/adapters/rhf'
+import { AdvancedFormStory } from '../_shared'
+
+const meta: Meta = {
+  title: 'Features/Form/React Hook Form/Advanced',
+  decorators: [
+    Story => (
+      <RHFAdapter>
+        <Story />
+      </RHFAdapter>
+    ),
+  ],
+}
+export default meta
+type Story = StoryObj
+
+export const Default: Story = {
+  render: () => <AdvancedFormStory />,
+}
